cosiap/cosiap_api/requirements.txt
django>=5.0.6,<5.1.0
djangorestframework>=3.15.1,<3.16.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
drf-spectacular>=0.27.2,<0.28.0
django-cors-headers>=4.3.1,<4.4.0
pillow>=10.3.0,<10.4.0
django-filter>=24.2,<24.3
django-model-utils>=4.5.1,<4.6.0
mysqlclient>=2.2.4,<2.3.0
django-environ>=0.11.2,<0.12.0
six>=1.16.0,<1.17.0

django-crontab>=0.7.1,<0.8.0

python-docx==0.8.11

channels[daphne]>=4.1.0,<4.2.0
channels-redis>=4.2.0,<4.3.0

gunicorn>=22.0.0,<22.1.0
uvicorn[standard]>=0.30.1,<0.31.0


